Add editor-translator as testable variable
Add editor-translator to the RNC schema to make it a testable variable. This makes it possible to modify the formatting of editor and translator depending on if they are the same person or not. The variable itself is still populated by the processor collapsing editor and translator.

This would require a minor update to processor code. They already have the editortranslator collapsing code; but they would likely need a small modification to call it during and if test. Given that, probably v1.1, rather than 1.0.2?

Fixes part of # citation-style-language/csl-evolution#7