Restructuring and Updating CONTRIBUTING.md (#312)
Alines with the terms from the README.md (e.g. CSL-JSON Schema or 
add csl-categories.rnc).

Also makes explicit a section about issues (with the text which 
was already there) and moved this together with PR on the top. 

The rules for versioning are much less frequent and therefore 
following, but also renamed from their previous title.