input: Add experimental rich-text support
This adds an experimental csl-rich-text.yaml schema that defines a
structure for rich-text formatting in JSON.

Also adds a definition for rich-text variables, and title-string
definitions that uses that variable, and then redefines all title and
other fields to use these definitions. So it will be easy to merge the
rich text support in the future.

Addresses in part #278