Add `label` attribute to names schema
Used to override the default label rendered for a role.

Following the discussion at https://github.com/citation-style-language/schema/issues/293 and https://forums.zotero.org/discussion/84042/tromso-recommendations-for-citation-of-research-data-in-linguistics this lets a user enter a more specific role label without requiring such things to have their own variable types. For example, a linguist might cite the person whose speech they are studying as an `author` with `label: "speaker"`.